💄 style: Move model and provider icon components to @lobehub/icons
