+++ package.json
@@ -119,7 +119,7 @@
     "@langchain/community": "^0.2.31",
     "@lobehub/chat-plugin-sdk": "^1.32.4",
     "@lobehub/chat-plugins-gateway": "^1.9.0",
-    "@lobehub/icons": "^1.28.0",
+    "@lobehub/icons": "^1.29.1",
     "@lobehub/tts": "^1.24.3",
     "@lobehub/ui": "^1.149.2",
     "@neondatabase/serverless": "^0.9.4",
