+++ src/app/(main)/settings/llm/ProviderList/Bedrock/index.tsx
@@ -1,9 +1,7 @@
 'use client';
 
-import { Aws, Bedrock } from '@lobehub/icons';
-import { Divider, Input, Select } from 'antd';
+import { Input, Select } from 'antd';
 import { useTranslation } from 'react-i18next';
-import { Flexbox } from 'react-layout-kit';
 
 import { BedrockProviderCard } from '@/config/modelProviders';
 import { GlobalLLMProviderKey } from '@/types/user/settings';
@@ -57,12 +55,5 @@ export const useBedrockProvider = (): ProviderItem => {
         name: [KeyVaultsConfigKey, providerKey, 'region'],
       },
     ],
-    title: (
-      <Flexbox align={'center'} gap={8} horizontal>
-        <Aws.Color size={32} />
-        <Divider style={{ margin: '0 4px' }} type={'vertical'} />
-        <Bedrock.Combine size={26} type={'color'} />
-      </Flexbox>
-    ),
   };
 };
