--- src/features/AgentSetting/AgentTTS/options.ts
@@ -1,16 +1,0 @@
-import { SelectProps } from 'antd';
-
-export const ttsOptions: SelectProps['options'] = [
-  {
-    label: 'OpenAI',
-    value: 'openai',
-  },
-  {
-    label: 'Edge Speech',
-    value: 'edge',
-  },
-  {
-    label: 'Microsoft Speech',
-    value: 'microsoft',
-  },
-];
