+++ src/features/AgentSetting/AgentTTS/options.tsx
@@ -1,0 +1,19 @@
+import { Azure, OpenAI } from '@lobehub/icons';
+import { SelectProps } from 'antd';
+
+import { LabelRenderer } from '@/components/ModelSelect';
+
+export const ttsOptions: SelectProps['options'] = [
+  {
+    label: <LabelRenderer Icon={OpenAI.Avatar} label={'OpenAI'} />,
+    value: 'openai',
+  },
+  {
+    label: <LabelRenderer Icon={Azure.Avatar} label={'Edge Speech'} />,
+    value: 'edge',
+  },
+  {
+    label: <LabelRenderer Icon={Azure.Avatar} label={'Microsoft Speech'} />,
+    value: 'microsoft',
+  },
+];
